imported from version 0.02, see Changes file.

+++ examples/release.sh
@@ -11,16 +11,21 @@ echo `date +"$date_format"` "START" >>logs/release.log
 
 perl bin/release.pl --config=data/release.ini
 
-echo `date +"$date_format"` "Compressing Uploads data..." >>logs/release.log
+echo `date +"$date_format"` "Compressing Release data..." >>logs/release.log
 
-cd $BASE/dbx
-rm -f release.*
-cp $BASE/release/data/release.db .  ; gzip  release.db
-cp $BASE/release/data/release.db .  ; bzip2 release.db
-cp $BASE/release/data/release.csv . ; gzip  release.csv
-cp $BASE/release/data/release.csv . ; bzip2 release.csv
+if [ -f $BASE/release/data/release.db ];
+then
 
-mkdir -p /var/www/cpandevel/release
-mv release.* /var/www/cpandevel/release
+  cd $BASE/dbx
+  rm -f release.*
+  cp $BASE/release/data/release.db .  ; gzip  release.db
+  cp $BASE/release/data/release.db .  ; bzip2 release.db
+  cp $BASE/release/data/release.csv . ; gzip  release.csv
+  cp $BASE/release/data/release.csv . ; bzip2 release.csv
+
+  mkdir -p /var/www/cpandevel/release
+  mv release.* /var/www/cpandevel/release
+
+fi
 
 echo `date +"$date_format"` "STOP" >>logs/release.log
